requirements.txt
streamlit
pdfplumber
pytesseract
pdf2image
python-docx
Pillow
